Fix for flaky test test_chunks_the_same
This would fail sometimes on Windows / Mac because the clocks used there are less precise and the two [1000, 1001] segments would collide on exactly the same key with LMDB. I just change the test so their content hashes would be different. The test was written to cover an issue with identical index ranges, so it still meets its original purpose.

This just changes the tests so that their content hashes differ. Theoretically, this could be a real problem that comes up, and we ought to somehow retry when there is a key clash. But that seems very improbable outside of unit testing, so we leave it as is for now.