Get object sizes based on S3's ListObjects output (just with scan_object_sizes() to start with) (#2248)
Monday: 8560764974

Limited to just `scan_object_sizes` to start with, to show the idea.

Calculate object sizes using functionality in the storage backend itself
when possible. This PR calculates compressed sizes on S3 based on the
`ListObjectsV2` output, so we don't need to read all the keys.

For storages where we haven't implemented anything fancy, just read all
the (compressed) keys and check the compressed size on their header.

Remaining work:

- Use this approach for `scan_object_sizes_by_stream`
- Make sure we have testing against all backends especially NFS
- Implement `scan_object_sizes_for_stream`
- Native size calculations for LMDB and Azure
- Native size calculation for library size (LMDB has an API to do this
"in one")
- `AdminTools` Python API on the v2 API, including search by regex. Docs
page to explain the output.