Correctly handle `zoneinfo` timezones
Fixes #2245

Correctly normalizes zoneinfo timezones. Currently if we write something
as a pytz or a ZoneInfo timezone we'll read it back as a `pytz` to
preserve old pytz behavior and not break old arctic readers.

Consensus is that people generally prefer `ZoneInfo` and we can change
to always using `ZoneInfo` but we'll need to do so in next major
release.