Simple API to remove several staged symbols at once.
Private in NativeVersionStore for now so we don't need to give API guarantees till we know whether it's useful.

The proper "batch" API where we give back `Union[VersionedItem, DataError]` is nicer - and we'll have to do it if we promote to the Library API - but would need changes to all the storage backends.