Bugfix 8873431775: Fix NFS-backed storage sharding when deleting objects (#2301)
#### Reference Issues/PRs
Fixes 8873431775

#### What does this implement or fix?
A regression was introduced in
https://github.com/man-group/ArcticDB/pull/2176, specifically the change
in `nfs_backed_storage.cpp`, that resulted in the shard directory being
calculated incorrectly for NFS-backed storages when deleting objects.
This manifested as failing to delete snapshots, which is particularly
noticeable because they then still appear in the output of
`list_snapshots`. This would likely have resulted in no data being
deleted correctly from NFS-backed storages, but for other key types this
would present as orphaned data rather than errors, as data being
logically deleted is represented by a write operation to the version
chain.