Bugfix 9123099670: fix resampling of old updated data (#2351)
#### Reference Issues/PRs
Fixes
[9123099670](https://man312219.monday.com/boards/7852509418/views/168855452/pulses/9123099670)

#### What does this implement or fix?
Fixes three separate resampling bugs:

1. Old versions of `update` (changed sometime between `4.1.0` and
`4.4.0`, I haven't pinned down exactly where) had a behaviour in which
the `end_index` value in the data key of the segment overlapping with
the start of the date range provided to the `update` call was set to the
first value of the date range in the `update` call. For all other
modification methods, this is set to 1 nanosecond larger than the last
index value in the contained segment. Resampling assumed this to be the
case, and had an assertion verifying it. Relaxing this assertion is
sufficient to fix the issue.
2. Providing a `date_range` argument with a resample where the provided
date range did not overlap with the timerange covered by the index of
the symbol led to trying to reserve a vector with a negative size. This
now correctly returns an empty result.
3. Previously, checks that a symbol being resampled had a timestamp
index occurred after some operations which also require this to be true,
which could lead to the same vector reserve issue above. It is now
checked in advance, and a suitable exception raised.