Update recursive normalizers layout assertions
This is to reflect PR #2352. Looking at its `version_map.hpp` change, we change from writing two version keys when we write and tombstone_all (one for the tombstone, one for the write) to folding them together in to a single version key. This has reduced the number of version keys we write by one.