Bugfix/9256801170/memory over allocation when reading shorts dfs (#2404)
#### Reference Issues/PRs

[9256801170](https://man312219.monday.com/boards/7852509418/pulses/9256801170)

#### What does this implement or fix?
When presizing a `ChunkedBuffer`, the behaviour was to round this up to
the default block size (3968 bytes) if fewer bytes were requested,
resulting in a major over-allocation if the buffer would only ever hold
a few values.
Inspecting the code makes it clear that this method is only ever used
with buffers that will never subsequently be expanded, and so there is
no benefit to this overallocation, and we should always create a block
that has exactly the correct buffer size.

The test output from the newly added tests (these will be backed out
prior to merging):
```
	test_read_short_sym
		pre-fix
			RSS: 849.41 MB
			RSS: 1370.93 MB
			RSS: 2002.95 MB
			RSS: 2525.48 MB
			RSS: 3158.48 MB
			RSS: 3736.20 MB
			RSS: 4258.29 MB
			RSS: 4834.31 MB
			RSS: 5468.00 MB
			RSS: 6100.84 MB
		post-fix
			RSS: 457.21 MB
			RSS: 586.88 MB
			RSS: 826.82 MB
			RSS: 956.86 MB
			RSS: 1197.72 MB
			RSS: 1383.61 MB
			RSS: 1513.94 MB
			RSS: 1697.97 MB
			RSS: 1939.73 MB
			RSS: 2180.68 MB
	test_read_long_sym_date_range
		pre-fix
			RSS: 1105.40 MB
			RSS: 1675.86 MB
			RSS: 2246.67 MB
			RSS: 2821.30 MB
			RSS: 3391.05 MB
			RSS: 3967.81 MB
			RSS: 4542.89 MB
			RSS: 5109.78 MB
			RSS: 5685.29 MB
			RSS: 6254.56 MB
		post-fix
			RSS: 714.96 MB
			RSS: 895.02 MB
			RSS: 1076.36 MB
			RSS: 1258.63 MB
			RSS: 1435.95 MB
			RSS: 1618.65 MB
			RSS: 1801.58 MB
			RSS: 1984.29 MB
			RSS: 2166.22 MB
			RSS: 2344.72 MB
```

#### Any other comments?
Even with this change, we are still holding onto a lot of memory in
addition to the raw buffers required to back the numpy arrays. A more
thorough fix will come as part of
[9106978040](https://man312219.monday.com/boards/7852509418/pulses/9106978040)
once the initial Arrow reading PR is merged