Validate staged data symbol names (#2414)
We have quite strict validation we apply to newly written symbols at the
time of the write. But this logic is accidentally not implemented for
staged writes, where we should blow up when we compact and write the new
symbol. See `name_validation.cpp` for our validation logic - symbols
must be at most 254 chars long, in ASCII range [32, 126] and must not
include `*` `<` or `>`.

If a symbol already exists with a name that fails this validation we
still allow staged writes to it - to keep workflows that predate our
stricter validation working.

See Monday ticket 8644487034 for more.