Read as arrow backend (#1831)
Monday ref: 7855121013

Backend work for reading arrow. This PR should not change existing
behavior in any way. Fully exposing reading as arrow to users will
happen in a separate frontend PR.
Currently reading as arrow is only allowed on V1 API behind
`_output_format` kwarg (only to be used for tests).

Arrow layout description:
When reading a symbol we produce a `pyarrow.Table` which consists of
several `RecordBatch`es. The record batches correspond to our row sliced
segments. We might want to add different layout options in the future.
String columns are returned very similar to our internal representation.
They are returned as a `DictArray` from `int32` index in a
`BigStringArray`. The difference is that arrow requires the string
buffers to be part of the column unlike our `StringPool` which is shared
between columns. The `ArrowStringHandler` handles this conversion.
Note that we use `int32` instead of `int64` dict keys. That should be
fine because of the row group separation based on our data keys. We use
`int64` string offsets to allow usage for longer strings.

Changes included in this PR:
- All sparrow code required to zero-copy construct arrow arrays from
`SegmentInMemory`s. (contained in `cpp/arcticdb/arrow`)
- Adds a new Chunked buffer `AllocationType::DETACHABLE`. This allows
detaching the memory blocks and moving the memory to sparrow. (mostly
introduced in previous PR but refined in this one)
- Adds extra buffer hashtable to `Column`s. This is used to attach
additional `ExtraBufferType::OFFSET` and `ExtraBufferType::STRING`
buffers to specific column offsets. We also use
`ExtraBufferType::BITMAP` for all sparse columns. These extra buffers
will be used to zero-copy construct a dict string arrow array.
- Modifies the interface of type handlers to allow the type handler to
insert the extra buffers when processing arrow strings
- Adds a `UTF_DYNAMIC32` type. It will only be used when outputting
arrow (we should never save a column with such type)
- Applies `ColumnTruncation` logic to `OutputFormat::Arrow`. This way
arrow output doesn't require post-processing. (contained in
`read_frame.cpp` and chunked buffer logic). The logic is partially
broken from previous PRs. I will write the fixes and more thorough
testing in a separate PR.
- `ReadResult` change to include `output_format` and have a
`variant<PandasOutputFrame, ArrowOutputFrame>`.


Commits:
The PR consists of two commits. The first one is squashed from William's
initial work on the branch. It consists of the majority of the changes.
The second one is including mostly bugfixes and refactors to make review
easier.
Note to reviewers: It would be best to review the commits together
unless you're already familiar with the first commit.


Known issues to be fixed in later PRs:
- No tests for arrow sparse data (+ several known issues with current
sparse data marked with TODOs in this PR)
- Fixed string handling with `OutputFormat::Arrow`
- Truncation is very broken. Will be rewritten in a follow up PR