Validate snapshot names  (#2415)
We have quite strict validation we apply to newly written symbols at the
time of the write. But this logic was accidentally not implemented for
snapshots. See `name_validation.cpp` for our validation logic - symbols
must be at most 254 chars long, in ASCII range [32, 126] and must not
include * < or >. This PR introduces this validation for snapshot names.

Monday: 8667974441

Similar to PR 2414 for staged data.