Fix update slices selection
Due to the changes in previous commits to only load the required date
range update slices selection was broken because it was relying on off
by 1 inclusive/exclusive errors.

This commit fixes the issues with update and adds a lot of query stats
test which verify we do exactly as many reads as we need to for each of
the operations.