Performance 9380719845: Reduce memory overhead when reading dataframes from ArcticDB (#2435)
#### Reference Issues/PRs

[9106978040](https://man312219.monday.com/boards/7852509418/pulses/9106978040)

#### What does this implement or fix?
Prior to this change:

1. The `SegmentInMemory` returned by calls to `read` (and similar) was
wrapped in a Python object, which was then used as the base object for
each individual `ndarray`. This meant that even if 99% of the arrays
were freed, the entire `SegmentInMemory` (i.e. all columns) would not be
freed until the final array was freed.
2. The buffers underlying this `SegmentInMemory` were allocated as
`presized`. In practice, this meant for small buffers that there was a
lot of overhead taken up with metadata and page-alignment.

These issues are addressed here by:

1. Giving each `ndarray` its own base object (the new class
`NumpyBufferHolder`) referencing only the memory needed for that
individual array.
2. Using the new (from Arrow work) `detachable` allocation type for
buffers that will eventually be used to back `ndarrays`. This removes
the page alignment, and allows all of the metadata around our buffer
class to be freed, leaving just the information necessary to decref any
objects necessary when the `ndarray` is destroyed, and then free the
memory.

This has also eliminated the need for ping-ponging off the Python layer,
and massively simplified the memory management of `PandasOutputFrame`
and `FrameData`, and completely removing `FrameDataWrapper`. The model
for Pandas is now almost identical to that with Arrow.

Change tested by writing data to an S3-backed library using:
```
def write_data():
    lib.version_store.clear()
    num_rows = 1_000
    num_columns = 10_000
    df = pd.DataFrame(np.random.randint(0, 100, size=(num_rows, num_columns)), index=pd.date_range("2025-01-01", periods=num_rows))
    lib.write("sym", df)
```
and then reading it back using:
```
def read_data():
    date_range = (pd.Timestamp("2025-01-01"), pd.Timestamp("2025-01-04"))
    retrieved = []

    for i in range(10):
        small = lib.read("sym", query_builder=QueryBuilder().date_range(date_range)).data
        retrieved.append(small)
        print(f"RSS: {psutil.Process(os.getpid()).memory_info().rss / 1e6} MB")
```
Results:
```
5.8.2  5.9.0  Branch
  345    313     313
  395    321     314
  443    330     315
  492    339     316
  539    346     318
  588    355     318
  636    362     320
  685    371     321
  734    380     322
  785    388     322
```
The difference between 5.8.2 and 5.9.0 is down to
https://github.com/man-group/ArcticDB/pull/2404

---------

Co-authored-by: Ivo Dilov <iddilov@gmail.com>