Fix some obfuscation that shouldn't happen when an index and a column share the same name
Bug fix for the column and index names returned by `get_description` and `get_info` APIs.

This is to fix
https://man312219.monday.com/boards/7852509418/views/168855452/pulses/8667920777
which was actually working correctly - the test case was incorrect (it
was not actually setting names on the index being written).

The testing exposes several rough edges with the index names exposed by
`get_info` on `Series` - but it does not seem particularly worthwhile
changing their behaviour now.