Fix flaky hypothesis tests for resampling
This is due to a bug in Pandas https://github.com/pandas-dev/pandas/issues/44957
When none of the values fall in any bucket Pandas, the groups in the resampler are empty, and Pandas
throws ValueError. ArcticDB behaves reasonably and returns an empty DataFrame.
This seems possible only if the origin is end_day