Enhancement 9402975662: Throw a meaningful exception when trying to read ASCII or fixed-width string columns into Arrow (#2529)
#### Reference Issues/PRs

[9402975662](https://man312219.monday.com/boards/7852509418/pulses/9402975662)

#### What does this implement or fix?
Full support for fixed-width (and possibly ASCII) strings will be added
later, for now just give a meaningful error if they are read with the
output format specified as Arrow.
For context, ASCII strings were only available for writing in Python 2.7
builds of the internal Man Group ArcticDB project. Likewise, fixed-width
strings have not been the default for a very long time.