Update cibuildwheel due to windows CI issues (#2563)
We need the latest config in this file
https://github.com/pypa/cibuildwheel/blob/v3.1.3/cibuildwheel/resources/virtualenv.toml
.

Older ci build wheel versions point at a dead link for get-virtualenv.

The new version requires Python >= 3.11 to launch, so I've changed the
script to use Python 3.13 to launch cibuildwheel. This doesn't affect
our actual builds which are still done in a manylinux container - only
the way that we launch them.