Performance/9697143845/mvp arrow performance profiling and optimisation (#2573)
#### Reference Issues/PRs

[9697143845](https://man312219.monday.com/boards/7852509418/pulses/9697143845)

#### What does this implement or fix?
Adds basic ASV benchmarks for Arrow reads, and improves the performance
of reading strings.
Eventually we can parametrize all the benchmarks with Arrow as the
input/output formats, this just makes sure we don't break anything
egregiously until we get to that point.

#### Benchmarking findings summary
Timings:

- Plain read calls `lib.read(sym)`
- Numeric data. Tested with 10 columns/1, 1k, 100k, 1m, and 100m rows of
int64 data
- Arrow performance identical to performance returning Pandas when
consolidation is disabled (the default with the V2 API for Pandas
>=2.0.0). This is expected, as the only real difference in this case is
the output frame is made up of multiple buffers per column with Arrow,
and just one with Pandas, but the timing difference of this is
negligible, and the time is dominated by decoding + memcpying with both.
- Pandas with consolidation enabled (the default inside Man) is strictly
slower, although this not noticeable at very low row counts.
- String data. Tested with 10 columns/1, 100, 10k, and 1m rows/1, 100,
and 100k unique 10 character strings
- No noticeable difference with <=100 rows as timings dominated by other
operations
- On a single core, Arrow ~x2 faster than Pandas before this change, and
~x3 faster after it
        - Using more cores increased the performance difference
- Date range read calls `lib.read(sym, date_range=blah)`
- Tested with a variety of ranges to test the various Arrow truncation
code paths and Pandas post-processing
    - Numeric data. Tested with 10 columns/1m rows of int64 data
        - Additional memcpys are visible in timings
- In the pathological case of excluding just 1 row from each end of the
symbol, adds ~17% compared to Pandas with consolidation disabled, taking
performance back to about where Pandas is with consolidation enabled
- String data. Tested with 10 columns/1m rows/1, 100, and 100k unique 10
character strings
        - Same performance hit observed as with numeric data
- But because the strings performance is faster than Pandas, overall it
is still always faster to use Arrow

Peak memory usage:

- For reading 10 columns x 100m rows = 1B 64 bit integers (8GB of data),
Arrow needs a few hundred MB in addition to the raw data size, and
Pandas without consolidation is similar. Pandas with consolidation is
about double as it memcpys all of the buffers we return into one larger
buffer
- Reading 10 columns x 1m rows = 10m strings with 1/100/100k unique
strings
    - Arrow needs 400MB/400MB/666MB
    - Pandas without consolidation needs 400MB/400MB/900MB
    - Pandas with consolidation needs 500MB/500MB/1GB