Fix hypothesis tests that occasionally make the CI runners OOM
The issue is that we need to output a dense dataframe for Pandas. So if
the date range is big and the frequency is small we can end up with too
many rows in the output dataframe eating the memory of the runner.

This also allows us to test all possible frequencies and extend the date
range for generating dataframes.