Enhancement 8987170104: arrow write numeric dense data (#2663)
#### Reference Issues/PRs

[8987170104](https://man312219.monday.com/boards/7852509418/pulses/8987170104)

#### What does this implement or fix?
Allows writing of `pyarrow.Table` objects provided they contain only
dense numeric data. This is by no means the final API and so must
currently be opted into by calling `_set_allow_arrow_input()` on
`NativeVersionStore`. Everything is zero copy as far as possible, with
the unavoidable exception of bool columns, as our internal
representation uses 1 byte per value, whereas Arrow uses 1 bit per
value.

Performance is pretty much identical to Pandas in comparable use cases,
with the exception of bool columns for the reason stated above, and
wide, shallow dataframes, which are a bit faster to write with Arrow
(presumably because of normalization stuff happening in the Python layer
with Pandas that we can skip with Arrow).

#### Any other comments?
In addition to not supporting strings or sparse data, there are still a
lot of improvements to be made:

- `update` with a `date_range` overlapping the provided table index not
currently supported
- Converting from Arrow's packed bitset bool column representation to
our internal format is not parallelised
- Non-nanosecond timestamp columns can be written, but are returned as
nanosecond timestamp columns
- Non-UTC timestamp columns also not handled correctly
- The index column (if specified) is only guaranteed to be at the
correct index when reading all data, not when selecting specific columns
or performing processing
- Only `pyarrow.Table` is supported right now, not lower level
primitives such as `RecordBatch/Array/ChunkedArray`
- On read, processing pipeline operations on data written as Arrow
probably won't work
- `get_info/get_description` output will not be accurate for data
written as Arrow
- No testing of appending/updating Arrow data with Pandas data and
vice-versa, so probably doesn't work
- Index column must always be specified for timeseries data, even when
it could be inferred in `append/update` calls