Performance 18146257858: speedup reading as_of tombstoned versions alive in snapshots (#2699)
#### Reference Issues/PRs

[18146257858](https://man312219.monday.com/boards/7852509418/pulses/18146257858)

#### What does this implement or fix?
Speeds up reading as_of a specific version number where the version
number has been deleted, but is alive in at least one snapshot (note
that this is only supported in the V1 API, we made a deliberate choice
for this to throw a NoSuchVersionException in the V2 API as the
performance can be so bad).

The main performance improvement comes from parallelising and
short-circuiting the reading of snapshot ref keys read from storage. The
previous implementation read all of the snapshot ref keys, and then
searched through for index keys with the correct symbol and version ids.
The new implementation reads snapshot ref keys in parallel, and stops
reading snapshot ref keys once it finds an index key with the correct
symbol and version id.

There are also some minor improvements to only materialise the atom key
in memory of the desired symbol/version pair, rather than all index keys
from all snapshots with the correct symbol, although in practice this
should be dwarfed by the IO improvements.