Remove `DataTypeMode` and `TypeHandler::type_size`
`DataTypeMode` was used to signify whether we are constructing the
segment to be offloaded to python or to be used internally. However it
was not assigned correctly and was not really used in a meaningfull way.

Also `TypeHandler::type_size` is not needed as the type size can be
inferred from the column type.