Enhancement 9722907222+18364157684: Denormalize all supported Pandas index and column names into Arrow (#2759)
#### Reference Issues/PRs

[9722907222](https://man312219.monday.com/boards/7852509418/pulses/9722907222)

[18364157684](https://man312219.monday.com/boards/7852509418/pulses/18364157684)

#### What does this implement or fix?
Pandas is very permissive in terms of it's column naming, allowing
arbitrary objects as column names, as well as duplicates, and no names
at all. This also applies to index columns. Of this large universe of
possible column names, we allow:

- Strings (including the empty string "")
- Integers
- `None` (== no name)
- Duplicates

Arrow is much more restrictive, requiring all column names to be unique
string values. As a large amount of data has already been written with
these quirky column names, we want to be able to read all of it back in
as Arrow format, with some kind of sensible naming scheme.

#### Prior Art
`pyarrow.Table.from_pandas` will use a stringified version of integer
and `None` column names. Unnamed index columns are named
`__index_level_i__`, where `i` is the level of the index, and placed as
the last columns in the table, and seems to have a bug where if there is
already a column with this name, then there will be duplicate column
names. Duplicate column names will raise a `ValueError`.

`polars.from_pandas` will also stringify integer and `None` column
names. Polars does not include index columns from Pandas by default,
`include_index=True` must be passed, in which case the columns will be
placed as the first columns in the resulting dataframe. Unnamed index
columns will be called `"None"`. If a multiindex is provided with
multiple unnamed columns then a `ValueError` is raised. Duplicate column
names will also raise a `ValueError`.

#### Design
We must be more permissive than `pyarrow` or `polars` due to the data
already written with these column names. The name resolution process is
as follows:

- Existing names are always respected as far as possible
- If the data has a single unnamed timeseries index, this will be named
`__index__` and placed as the first column in the resulting dataframe.
- If the data has a multiindex with one or more unnamed levels, these
will be named `__index_level_i__` where `i` is the level of the index
without a name, and placed as the first columns in the resulting
dataframe.
- Integer and `None` column names will be stringified.
- Duplicate column names will be handled by prepending and appending an
underscore character until a unique name is achieved. e.g. column names
`["hello", "hello", "hello"]`, would become `["hello", "_hello_",
"__hello__"]`
- This also applies to integer and `None` column names if they are
repeated.
- This also applies to unnamed index columns. e.g. if there is already a
column called `__index__` (with 2 underscores), then the index column
will become `___index___` (with 3 underscores).

In this way we can read back all data we have previously written, and it
will generally be possible to infer what the original column name was.