Performance 10588043604: Decode only required columns in processing pipeline with dynamic schema (#2766)
#### Reference Issues/PRs

[10588043604](https://man312219.monday.com/boards/7852509418/pulses/10588043604)

#### What does this implement or fix?
`PipelineContext::overall_column_bitset_` is used to construct a hash
set of columns that are to be decoded in the processing pipeline path.
Prior to this change, this bitset was only populated for static schema
and (unsupported) bucketize dynamic schema data. This resulted in all
columns being decoded with dynamic schema, which becomes a performance
bottleneck with very wide dataframes where only a few columns are
required for the processing and output.

Test coverage already fairly comprehensive in
`test_basic_version_store.py::test_dynamic_schema_read_columns`, and
`read_index` behaviour in `test_read_index.py`.