Use forwad declarations for SegmentInMemory and Column
These are one of the heaviest files in our codebase. Thus it makes sense to include them in as little header files as possible. This PR does not change any functionality it only moves some code using SegmentInMemory or Column from header files to .cpp and relacing the include with a forward declaration where possible.