Enhancement 18352299908: Allow configuring multiindex string column Arrow return format (#2788)
#### Reference Issues/PRs

[18352299908](https://man312219.monday.com/boards/7852509418/pulses/18352299908)

#### What does this implement or fix?
The Arrow string return type for named multiindex levels can now be set
with the `arrow_string_format_per_column` override.

There is an edge case explained in the comment in `arrow_handlers.cpp`
where a column the user has explicitly named `__idx__blah` can get the
wrong output format if there is also a column called `blah` with an
override return type set, but this is much more painful to fix and seems
unlikely to occur in practice.
[Ticket](https://man312219.monday.com/boards/7852509418/pulses/10679807500)
opened for tracking purposes just in case.