Enhancement 10681189884: Upgrade sparrow to 1.4.0 (#2789)
#### Reference Issues/PRs

[10681189884](https://man312219.monday.com/boards/7852509418/pulses/10681189884)

#### What does this implement or fix?
Updates sparrow to 1.4.0. This requires a vcpkg baseline update, which
does not work out of the box on linux. Binary search of vcpkg commits
identifies [this](https://github.com/microsoft/vcpkg/pull/44653) as the
culprit. We do not seem to be the only project affected
([ticket](https://github.com/microsoft/vcpkg/issues/48496)).
This only breaks the build of the arrow dependency, which is only used
for testing decimals, which are not yet supported, so for now I've just
removed the dependency and corresponding tests.