Allow fortran styled columns starting with None and NaN
Up to now exception was thrown. We can iterate through the column and use the stride to jump to the next element. Note that if the column does not start with NaN or None it's stored correctly because a flattener is applied. The exception is thrown by the code checking the type.