Benchmarks for MemBlock
MemBlock refactor

This is only a refactor to prepare for parallel bool unpacking and
should not change existing behavior.

It splits up the `MemBlock` into two separate classes inheriting a common
interface `IMemBlock`:
- `DynamicMemBlock` is responsible for the `AllocationType::DYNAMIC` and
  has a capacity which can gradually be filled up with resizes. This is
the type of memory blocks we use for in memory operations with
`ChunkedBuffer`s e.g. when tick streaming or when performing appends
etc.
- `ExternalMemBlock` is responsible for the `AllocationType::DETACHABLE`
  and for external pointers to views. It just holds a pointer to
external memory which can be owning or not. This is what we use when
either when returning Segments to python or when receiving views from
python

This change also adds the scaffolding needed for the packed bits buffer
via `ExternalPackedMemBlock`. It will be used in a follow up PR to
parallelise bool unpacking.

Address Alex's comments

Address Vasil's comments

Switch low level chunked_buffer checks to debug

The checks in particular getting the `physical_bytes` introduced a
slowdown in `BM_iteration_via_scalar_at` because:
- Previously `bytes()` call could be inlined and it can no longer be
  inlined because of the runtime polymorphism
- The non inlined function calls are taking a big part of the time.

Removing the check in `operator[]` removes two function calls in release
builds. This brings performance roughly inline to before the refactor.