Speed up the resample benchmark run (#2831)
It is dominating benchmarking on PRs at the moment.

After we turned these back on, the benchmarking time on PRs jumped from
about 2h20m to about 4h30m.

I followed the approach here
https://github.com/man-group/ArcticDB/wiki/Dev:-ASV-Benchmarks#flakiness
to verify that the benchmark is stable with the new parameters, and an
order of magnitude faster.